Flaskapp/requirements.txt
Flask
Tensorflow
Pillow
gradio
numpy